consolidate shared player/enemy functionality, fix enemy move direction
Enemies and Players were doing largely the same things for movement, so I pushed that down to a GameActor that they now both use. Additionally enemies were moving in circles around the player instead of just heading straight to them, and it appears to have been caused by me normalizing the vector from the enemy to the player. It's already clamped before we move, so no need to also normalize it.