add custom itch publish action, example usage
not ready to turn this on yet since i've had enough trouble getting it setup and it's tough to test fully.

the existing butler action requires creating a container as part of the action, and only linux runners support that, apparently. there are pre-built butler binaries available, but none for arm. so we pull and build butler ourselves.