fix jaggies
apparently my desktop pc's default pixel ratio is 1.5, so me setting it to 1 was overriding it. i thought 1 was the default so setting it explicitly was harmless, but i was wrong. so then i removed a bunch of other stuff that is the default so as to not accidentally interfere with settings, and now everything looks better. go figure.