switch to custom pause implementation
we are going to need to use Pause on level up to present the player with upgrade choices. the hack-pause i was doing before comes with too many side effects. this is really ugly and requires a lot of hooks and tracking, but i'm not sure we have any other options with Excalibur

on the plus side, Start on the gamepad can pause now. so there's that.