move game rendering to a Canvas
this gives us much better control over the html ui and allows the html to be setup the way you'd expect. previously nothing would show over the game unless its position was absolute, which busted a bunch of attempts at layouts of various hud elements.

not all of the html has been converted yet to be compatible (namely the paused/died text is drawing too low since they're leaving room for the ui-root div above them) but this is the base set of changes required to get something to build off of.

also note the gift.ts scale changes are because of a bug, as explained in the comment in the code. hopefully it's fixed in the engine soon, but this workaround is fine for now.