clean up multi-amount spawns
now there's a rotational or temporal variance (based on the type of weapon) when spawning multiple of a weapon. i tried to get the orbit timings working such that they'd spawn with an even distribution, and it...kinda works...at least until you start upgrading the speed of the orbiter

this also fixes a bug where seeker weapons could get stuck somewhere with a zero direction/movement vector.