fix enemies not taking damage from every hit
i forgot the invulnerability window was a base GameActor concept. it was only ever supposed to apply to players, so now it's opt-in rather than opt-out.

this was causing all sorts of weird breakage like enemies not dying when they hit the player if they had taken damage within the invuln window prior to hitting the player

:facepalm: