add perlin noise camera strategy
this provides the ability to translate the camera position using noise for a smooth and predictable camera motion.

note that a lot of these units are somewhat arbitrarily chosen to get something that felt "right", and as a result the inputs are somewhat meaningless in terms of scale. if this weren't a game jam, i'd spend more time trying to get coherency here, but alas...