fix high-amount weapon spawning
the actions component was too hard to use for this purpose, so we now manage the multi-spawns ourselves, with delays scaled by the number of weapons we need to spawn and a limit placed on how many can be spawned (if we want to spawn more than we have the time to between now and when the next interval pops, just drop all the extras). this means that +amount stops functioning at a certain -interval, but you have to level up a ton and/or get really lucky upgrade rolls for that to happen, so for now i'm going to try this out.