fix sample bleeding between tiles in some cases
after asking on the excalibur discord, it seems that uvPadding exists to alleviate this issue. uvPadding is set to 0.25 automatically when pixelArt is enabled, but pixelArt by itself was causing some ugly seams between tiles already; it appears that the AA settings i have right now fix that, though, so we can turn pixelArt back on while keeping AA off and all seems well with the world.