initial spawn timings refactor attempt
this is probably a little too overtuned on the hard side, but it is a starting point for a new spawning methodology. now we aim to keep a minimum number of enemies alive at any given time, and check every second to ensure that's the case. we can define these numbers per-wave and define how long a given wave lasts plus what enemies are eligible to spawn in a wave.