flesh out enemy spawns a bit more, flatten speeds
too many multipliers were being applied to make speed a consistent-scalable value in the json, so this always takes the exact value from the json and applies it to the enemy.

i need either more enemies or stronger enemies (or both) later in the game, and it would probably be worthwhile to have some unique 'boss' spawns as well. but this is really coming together now!