always round upgrades down
i was shown that i was getting +2.0 amount, but then only 2 rockets were firing instead of the expected 3. inspecting in the debugger showed that my amount was 2.965, so the previous toPrecision call rounded the 1.965 up to 2.0. now we floor the result of multiplying the number by 10 and then divide back out before rendering a number with 1 decimal point, so we'll never show higher than we're actually getting, only lower